requirements.txt
biopython==1.78
h5py==2.10.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
numpy==1.18.5
PyVCF==0.6.8
scipy==1.5.2
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0